Moving files to move appropriate places. Removing skips for lint checks. Remvoing unused stylesheet. Removing extra imports.
