+++ web/src/__tests__/reducers/jobs.test.ts
@@ -1,7 +1,7 @@
 import * as actionTypes from '../../constants/ActionTypes'
 import _find from 'lodash/find'
 import _sample from 'lodash/sample'
-import jobsReducer, { initialState } from '../../reducers/jobs'
+import jobsReducer, { initialState } from '../../store/reducers/jobs'
 
 const jobs = require('../../../docker/db/data/jobs.json')
