+++ web/src/components/Dialog.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { dialogToggle } from '../actionCreators'
+import { dialogToggle } from '../store/actionCreators'
 import Button from '@material-ui/core/Button'
 import Dialog from '@material-ui/core/Dialog'
 import DialogActions from '@material-ui/core/DialogActions'
