+++ web/src/components/lineage/Lineage.tsx
@@ -6,14 +6,14 @@ import { DAGRE_CONFIG, INITIAL_TRANSFORM, NODE_SIZE } from './config'
 import { GraphEdge, Node as GraphNode, graphlib, layout } from 'dagre'
 import { HEADER_HEIGHT } from '../../helpers/theme'
 import { IDataset, IJob } from '../../types'
-import { IState } from '../../reducers'
+import { IState } from '../../store/reducers'
 import { MqNode } from './types'
 import { WithStyles, createStyles, withStyles } from '@material-ui/core/styles'
 import { Zoom } from '@visx/zoom'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 import { localPoint } from '@visx/event'
-import { setSelectedNode } from '../../actionCreators'
+import { setSelectedNode } from '../../store/actionCreators'
 import Edge from './components/edge/Edge'
 import Node from './components/node/Node'
 import ParentSize from '@visx/responsive/lib/components/ParentSize'
