+++ web/src/components/lineage/components/drag-bar/DragBar.tsx
@@ -4,7 +4,7 @@ import * as Redux from 'redux'
 import { Theme } from '@material-ui/core/styles/createMuiTheme'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { setBottomBarHeight } from '../../../../actionCreators'
+import { setBottomBarHeight } from '../../../../store/actionCreators'
 import Box from '@material-ui/core/Box'
 import classNames from 'classnames'
 import createStyles from '@material-ui/core/styles/createStyles'
