+++ web/src/components/search-bar/SearchBar.tsx
@@ -2,7 +2,7 @@ import * as Redux from 'redux'
 import { Theme, WithStyles, createStyles, fade } from '@material-ui/core/styles'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { findMatchingEntities } from '../../actionCreators'
+import { findMatchingEntities } from '../../store/actionCreators'
 import InputBase from '@material-ui/core/InputBase'
 import React from 'react'
 import SearchIcon from '@material-ui/icons/Search'
