+++ web/src/constants/ActionTypes.ts
@@ -3,18 +3,15 @@ export const DIALOG_TOGGLE = 'DIALOG_TOGGLE'
 
 export const FIND_MATCHING_ENTITIES = 'FIND_MATCHING_ENTITIES'
 
-export const FETCH_DATASETS = 'FETCH_DATASETS'
 export const FETCH_DATASETS_SUCCESS = 'FETCH_DATASETS_SUCCESS'
 export const FILTER_DATASETS = 'FILTER_DATASETS'
 
-export const FETCH_JOBS = 'FETCH_JOBS'
 export const FETCH_JOBS_SUCCESS = 'FETCH_JOBS_SUCCESS'
 export const FILTER_JOBS = 'FILTER_JOBS'
 
 export const FETCH_JOB_RUNS = 'JOB_RUNS'
 export const FETCH_JOB_RUNS_SUCCESS = 'FETCH_JOB_RUNS_SUCCESS'
 
-export const FETCH_NAMESPACES = 'FETCH_NAMESPACES'
 export const FETCH_NAMESPACES_SUCCESS = 'FETCH_NAMESPACES_SUCCESS'
 
 export const SET_SELECTED_NODE = 'SET_SELECTED_NODE'
