+++ web/src/globals.ts
@@ -6,3 +6,7 @@ declare const __API_URL__: string
 
 declare const __FEEDBACK_FORM_URL__: string
 declare const __API_DOCS_URL__: string
+
+export const API_URL = __API_URL__
+export const FEEDBACK_FORM_URL = __FEEDBACK_FORM_URL__
+export const API_DOCS_URL = __API_DOCS_URL__
