--- web/src/requests/namespaces.ts
@@ -1,8 +1,0 @@
-import { Namespaces } from '../types/api'
-import { genericFetchWrapper } from '.'
-
-export const fetchNamespaces = async () => {
-  // eslint-disable-next-line no-undef
-  const url = `${__API_URL__}/namespaces`
-  return genericFetchWrapper<Namespaces>(url, { method: 'GET' }, 'fetchNamespaces')
-}
