+++ web/src/store/reducers/display.ts
@@ -1,4 +1,4 @@
-import { APPLICATION_ERROR, DIALOG_TOGGLE } from '../constants/ActionTypes'
+import { APPLICATION_ERROR, DIALOG_TOGGLE } from '../../constants/ActionTypes'
 
 interface IToggleExpandAction {
   type: string
