+++ web/src/store/requests/namespaces.ts
@@ -1,0 +1,8 @@
+import { API_URL } from '../../globals'
+import { Namespaces } from '../../types/api'
+import { genericFetchWrapper } from './index'
+
+export const fetchNamespaces = async () => {
+  const url = `${API_URL}/namespaces`
+  return genericFetchWrapper<Namespaces>(url, { method: 'GET' }, 'fetchNamespaces')
+}
